moved quote search to a module. part of #495
